feat(api)!: seven owner rulings — widen recovery_percentage, and stop four blockers citing CI
Seven of the eleven rulings the owner made on 2026-09-02/03, plus the three
blocker restatements AD-0070 made possible. The remaining four (typed
content-total, multipart defect reporting, unreachable arms, and the v2-api
flip) follow separately.

RULING 6 — `recovery_percentage()` returns `Option<u16>`. This is the only
signature change in the batch. RAR 6.10 raised the recovery ceiling from 99%
to 1000% and moved the RAR5 encoding to a vint, so `rar -rr256p` upward
produces ordinary readable archives — and what the old signature did with them
was neither truncate nor error. The `u8::try_from` failed, `.ok()` collapsed it,
and the caller got `Ok(None)`, "the record is there, its percentage cannot be
determined", for a number sitting in plain sight in the header. The site was
already annotated with this ticket's own id: someone hit the limit, wrote down
the fix, and shipped the narrow type anyway.

Bounded rather than opened: the extra-area value is a `u64` vint, so past
`u16::MAX` stays on the `None` branch where 256 used to sit. RAR4 gains nothing
and the doc says so — it stores no percentage field at all, deriving the value
from block counts and capping at 100, so the widening is representational
there. `ReadArchive::recovery_percentage` moves with the facade.

RULING 1 — `ArchiveError::Password` documents what it cannot tell you. No code
change; a contract that was always true and never written down. On an encrypted
7z entry the variant means wrong password OR damaged payload, because 7z AES-256
carries neither an authentication tag nor a password-verification value: a wrong
key decrypts to plausible garbage that fails the downstream LZMA/CRC checks
exactly as damaged ciphertext does. Nothing is withheld — at that point the
distinction does not exist.

Scoped per backend rather than crate-wide, because the answer differs: RAR5's
per-file check value and ZIP's AES/ZipCrypto verification bytes do separate the
two; legacy RAR3 does not. The doc also warns that `validate_integrity` is not a
tiebreaker — it decodes the same ciphertext with the same key — which is the
thing a caller would most plausibly reach for. No typed ambiguity marker, and
that is the ruling: it would advertise a distinction the format cannot supply.
`ArchiveError` is `#[non_exhaustive]`, so adding one later is not breaking,
which is why documenting first costs nothing.

RULING 11 — DEF-005 is struck. Its exit criterion was "all remaining caveats
closed or formally accepted", and all three now are: encrypted-ZIP re-encryption
was already rejected; crash-recovery journaling is settled by the design
(tempfile + atomic rename, and `Limitations.md` says there is no journal); and
ZIP64 above 4 GiB is accepted as a permanent caveat rather than covered, since
multi-GiB fixtures are slow, awkward to keep in a suite, and too large to
commit. `Limitations.md` now says "accepted permanent caveat, not a gap awaiting
work", and tells a caller crossing that boundary to verify against an
independent tool.

RULINGS 3, 5, 9, 13 — tracker shape. ticgit deff990a folds into c1296744
(both break `src/entry.rs` in one release, and deff990a's own blocker asked for
exactly that — its two siblings closed without it, so the sequencing it wanted
is no longer achievable as written); OI-0076-003 splits, freeing two ready items
that were invisible behind two blocked ones, with 3f7dfa narrowed to items 5/6;
the libarchive in-place eager-probe gate is approved on 05e2dea4; and the OKF
migration is relabelled as tracked externally, so it stops reading as work
waiting on the owner.

THE THREE BLOCKER RESTATEMENTS, now that AD-0070 defines what CI means here:

  * OI-0065-001 — the blocker changes from "this repository has no CI
    configuration" to "this project has no Windows host". The first phrasing
    implied adopting a service would unblock it; under AD-0070 no service will
    be adopted. Required Action 4 asks for a recorded run rather than a CI job.
  * OI-0080-001 — stops citing CI at all. It is blocked by the standing pre-v2
    scope ruling, on its own, and always was. Recorded so the lane is not
    re-derived: the check half runs on this host today, and is exactly what
    would expose the host-cfg defect the issue names.
  * AD-0058 ordering constraint 2 — restated as release-gate lanes plus a
    verification record, and priced. Six *test* profiles would be two to three
    hours serial here; tests for the two profiles the constraint names plus
    `cargo check` for the other four protects the same property at roughly a
    quarter of the cost, and check lanes link no binaries so they cannot hit the
    first-exec admission stall.

Also corrected: CHANGELOG's `[Unreleased]` preamble claimed "nothing here
changes a signature". True when written, false as the window filled, and this
ruling makes it plainly false. Corrected rather than dropped, because a reader
who trusted it would have skipped the one break that does not announce itself at
the call site.

Gate, both profiles: check --all-targets --all-features 0; fmt 0; clippy
--all-targets --all-features -D warnings 0; clippy --no-default-features
--all-targets -D warnings 0; test --all-features 0 — 41 suites, 1980 passed;
test --no-default-features 0 — 41 suites, 1748 passed. Zero failures on both.

+++ src/archive.rs
@@ -2041,9 +2041,27 @@ impl Archive {
     /// - ZIP: ❌ Not supported (no recovery records)
     /// - TAR: ❌ Not supported (no recovery records)
     ///
+    /// # Range
+    /// `1..=1000` for RAR5. The percentage is stored as a vint since RAR
+    /// 6.10, which raised the maximum recovery record from 99% to 1000%, so
+    /// `rar -rr1000p` produces an ordinary readable archive whose percentage
+    /// does not fit a byte. This returned `Option<u8>` until 0.5.0 and
+    /// answered `None` for every such archive; widening it to `Option<u16>`
+    /// (ticgit 7ca208) is a breaking change taken inside the 0.5.0 window
+    /// rather than a major bump of its own.
+    ///
+    /// A RAR4 archive still cannot report more than 100: that format stores
+    /// no percentage field, so the value is derived from the recovery/total
+    /// block counts and capped. On the RAR4 path the wider type is
+    /// representational only.
+    ///
     /// # Returns
     /// - `Ok(Some(percentage))` - Recovery records present with known percentage
-    /// - `Ok(None)` - No recovery records present
+    /// - `Ok(None)` - No recovery records present, *or* the record is present
+    ///   but its size is not readable here — which is the answer for a
+    ///   header-encrypted (`rar -hp`) archive, whose percentage lives in a
+    ///   block this password-less byte walk is not permitted to decrypt
+    ///   (ticgit 3f8790). `has_recovery_record()` still reports `true` for it.
     /// - `Err(...)` - I/O error or parse error
     ///
     /// # Examples
@@ -2058,7 +2076,7 @@ impl Archive {
     /// }
     /// # Ok::<(), unified_archive::ArchiveError>(())
     /// ```
-    pub fn recovery_percentage(&self) -> Result<Option<u8>> {
+    pub fn recovery_percentage(&self) -> Result<Option<u16>> {
         // R0071-0012: same write-mode rejection as
         // `has_recovery_record` so a write handle never silently
         // answers `None`.
